requirements.txt
Django==2.0.5
djangorestframework==3.8.2
django-rest-swagger==2.2.0
gunicorn==19.8.1
django-cors-headers==2.2.0
django-filter==1.1.0
psycopg2==2.7.4
whitenoise==3.3.1